remove obsolete three-ar, add new webxr-ar
