+++ src/index.js
@@ -1,4 +1,5 @@
-require('./three-ar');
+require('./webxr-ar');
+//require('./three-ar');
 require('./mozilla-xr-ar');
 require('./ar-planes');
 require('./ar-anchors');
